requirements.txt
tensorflow
streamlit
pillow
numpy
pandas
matplotlib
seaborn
plotly
scikit-learn 